close shoppingcart ending? other fixes as well
yoink

+++ cantine-choupi/src/components/shoppingcart/shoppingcart.js
@@ -2,6 +2,7 @@ import React from 'react';
 import "./shoppingcart_styles.scss";
 import Button from '../button/button';
 import Call from '../../helpers/call';
+import { GoTrashcan } from 'react-icons/go';
 
 class Shoppingcart extends React.Component {
   constructor(props) {
@@ -45,19 +46,45 @@ class Shoppingcart extends React.Component {
     this.setState({
       items: items
     });
+
+    return window.updateCart(true);
   };
 
   // To decrease the amount of a product
   decreaseProductAmount(index_of_item) {
     let items = JSON.parse(window.localStorage.getItem('shoppingcart_items'));
     let item = items[index_of_item];
-    item.amount--;
 
-    window.localStorage.setItem('shoppingcart_items', JSON.stringify(items));
+    if (item.amount === 1) {
+      const confirmation = window.confirm("Are you sure?");
 
-    this.setState({
-      items: items
-    });
+      if (confirmation === true) {
+        const newItems = items.slice(index_of_item + 1, index_of_item + 2);
+        window.localStorage.setItem('shoppingcart_items', JSON.stringify(newItems));
+
+        const currentItems = JSON.parse(window.localStorage.getItem('shoppingcart_items'));
+        if (!currentItems) {
+          window.location.reload();
+        } else {
+          this.setState({
+            items: newItems
+          });
+
+          window.updateCart(true);
+        };
+      };
+    } else {
+      item.amount--;
+      window.localStorage.setItem('shoppingcart_items', JSON.stringify(items));
+
+      this.setState({
+        items: items
+      });
+    };
+
+    if (items.length === 0) {
+      this.clearCart();
+    }
   };
 
   getProducts() {
@@ -77,22 +104,51 @@ class Shoppingcart extends React.Component {
   getTotalAmount() {
     let amount = 0;
     let all = JSON.parse(window.localStorage.getItem('shoppingcart_items'));
-    all.map((item, index) => {
-      return amount += item.amount;
-    });
+    if (!all) {
+      return false;
+    } else {
+      all.map((item, index) => {
+        return amount += item.amount;
+      });
 
-    return amount;
-  }
+      return amount;
+    };
+  };
 
   getTotalPrice() {
     let total = 0;
     let all = JSON.parse(window.localStorage.getItem('shoppingcart_items'));
-    all.map((item, index) => {
-      let price = Number.parseFloat(item.product.price);
-      return total += (price * item.amount);
-    });
+    if (!all) {
+      return;
+    } else {
+      all.map((item, index) => {
+        let price = Number.parseFloat(item.product.price);
+        return total += (price * item.amount);
+      });
 
-    return Math.round((total + Number.EPSILON) * 100) / 100;
+      return Math.round((total + Number.EPSILON) * 100) / 100;
+    };
+  };
+
+  clearCart() {
+    window.localStorage.removeItem('shoppingcart_items');
+    window.updateCart(false);
+  };
+
+  getTotalProductPrice(index_of_item) {
+    let price;
+    let items = JSON.parse(window.localStorage.getItem('shoppingcart_items'));
+    let item = items[index_of_item];
+
+    price = (item.amount * item.product.price);
+    return Math.round((price + Number.EPSILON) * 100) / 100;
+  };
+
+  getTotalProductAmount(index_of_item) {
+    let items = JSON.parse(window.localStorage.getItem('shoppingcart_items'));
+    let item = items[index_of_item];
+
+    return item.amount;
   }
 
   render() {
@@ -118,7 +174,8 @@ class Shoppingcart extends React.Component {
                     <li className="clearfix">
                       <span className="item item-name">Name: <span className="pauper">{ item.product.name }</span></span>
                       <span className="item item-price">Price: <span className="pauper">{ `€${item.product.price}` }</span></span>
-                      <span className="item item-quantity">Quantity: <span className="pauper item-amount">{ item.amount }</span></span>
+                      <span className="item item-quantity">Quantity: <span className="pauper item-amount">{ this.getTotalProductAmount(index) }</span></span>
+                      <span className="item item-total-price lighter-text">Subtotal: <span className="pauper item-amount">€{ this.getTotalProductPrice(index) }</span></span>
                     </li>
                   </div>
                 )
@@ -126,6 +183,7 @@ class Shoppingcart extends React.Component {
             }
           </ul>
           <Button color="white">Checkout</Button>
+          <GoTrashcan className="remove_icon" onClick={ this.clearCart } />
         </div>
       </React.Fragment>
     );
